Refactor group extraction from Athena into separate tool #87
