--- larch_athena/macros.xml
@@ -1,1 +1,0 @@
-../common/macros.xml
