+++ larch_extract_group/test-data/ffi0.tabular
